requirements.txt
fastapi
pandas
numpy
plotly
uvicorn
psycopg2-binary
redis
alpaca-trade-api
streamlit